requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator
slowapi

# Testing
pytest
pytest-cov
httpx